Extract Header::SMB::Blocks to its own file.
